requirements.txt
fastapi
uvicorn
python-dotenv
ollama